CSVDATA

Description

Module that holds type constructors and methods for execption handling

Documentation

data Stock

Constructors

Stock

Fields

date :: String

open :: Double

high :: Double

low :: Double

close :: Double

volume :: Int

adj :: Double

key :: Int

data Company

Constructors

Company

Fields

identifier :: Int

name :: String

url :: String

link :: String

data InputError

Constructors

EmptyString

StringTooLong Int

WrongArg

NotInt String

OtherError String

ConnectionError String

AddPrompt

type InputMonad = Either InputError

InputMonad For our monad type constructor, we use Either InputError which represents failure using Left InputError or a successful result of type a using Right a. Reference : https://hackage.haskell.org/package/mtl-2.2.1/docs/Control-Monad-Error.html

convert :: String -> InputMonad Int

Converts an input id into an int, returns error otherwise Wraps input calculation to catch the errors. Returns either length of the string or an error.

check :: String -> InputMonad Int

Method to test whether the input is valid Attempts to calculate length and throws an error if the provided string is empty or longer than 3 characters. 3 was choosen because their will never be more than 999 companies in the database The processing is done in Either monad.

checkResult :: InputMonad Int -> Int

Method check whether result is correct, throws error otherwise

isN :: [Char] -> Bool

Method check whether an input is a number or not