CSVDB

Description

Method for handling Database connectivity and queries. | Some methods were used and modified according to Chapter 21 of 'Real World Haskell' | by Bryan O'Sullivan, Don Stewart, and John Goerzen. | Link :http:/book.realworldhaskell.orgread/extended-example-web-client-programming.html

Documentation

connect :: FilePath -> IO Connection

Initialise DB and return database Connection

prepDB :: IConnection conn => conn -> IO ()

Prepare the db for adding two tables.

two tables are created and data is verified * C_ID is unique

addStock :: IConnection conn => conn -> [Stock] -> IO ()

Adds many stock items to the database. Each item represents a line from the csv file. The entire csv file will be represented by many stock items. Each item also has a key, which represents the company it belongs to

addCompanies :: IConnection conn => conn -> [Company] -> IO ()

Adds many companies to the db Each company stores it's csv url and a link to it's html page

addCompany :: IConnection conn => conn -> Company -> IO Company

An attempt to add a new company to the db.

getCompanies :: IConnection conn => conn -> IO [Company]

Gets a list of all companies.

removeCompany :: IConnection conn => conn -> Int -> IO ()

Remove a Company. This also has to delete any stock which is assosiated with the company.

updateCompany :: IConnection conn => conn -> Int -> String -> IO ()

Updates an existing company. A more current csv link is added --However all the previous stock relating to the old link is deleted

getCompany :: IConnection conn => conn -> Int -> IO (Maybe Company)

Get a particular Company, by it's id.

getStock :: IConnection conn => conn -> Int -> IO [Stock]

Gets stock from one company. The ID is required to return the stock.

convStockRow :: [SqlValue] -> Stock

Converts a list of SqlValues to a Stock data type

convCompanyRow :: [SqlValue] -> Company

Converts a list of SqlValues to a Company data type

convStockToList :: Stock -> [SqlValue]

Converts a Stock data type to a list of SqlValues for insertion into the DB.

convCompToList :: Company -> [SqlValue]

Converts a Company data type to a list of SqlValues for insertion into the DB.

priceByComp :: IConnection conn => conn -> Int -> IO [[String]]

Method to print all price closes, sorted by price and for one company

priceByCompDateRange :: IConnection conn => conn -> Int -> String -> String -> IO [[String]]

Method to print all price closes, sorted by price and for one company

highestPrice :: IConnection conn => conn -> Int -> IO [[String]]

Method to print the highest price close, sorted by price and for one company

highestPriceDateRange :: IConnection conn => conn -> Int -> String -> String -> IO [[String]]

Method to print the highest price close, sorted by price and for one company

highestPriceAll :: IConnection conn => conn -> IO [[String]]

Method to print the highest price close, sorted by price and for all companies

highestPriceAllDateRange :: IConnection conn => conn -> String -> String -> IO [[String]]

Method to print the highest price close, sorted by price and for all companies

cheapestPriceAll :: IConnection conn => conn -> IO [[String]]

Method to print the lowest price close, sorted by price and for all companies

cheapestPriceAllDateRange :: IConnection conn => conn -> String -> String -> IO [[String]]

Method to print the lowest price close, sorted by price and for all companies

cheapestPrice :: IConnection conn => conn -> Int -> IO [[String]]

Method to print the cheapest price close, sorted by price and for one company

cheapestPriceDateRange :: IConnection conn => conn -> Int -> String -> String -> IO [[String]]

Method to print the cheapest price close, sorted by price and for one company

avgPriceAll :: IConnection conn => conn -> IO [[String]]

Method to print the avg price closes, sorted by price and for all companies

avgPriceDateRange :: IConnection conn => conn -> String -> String -> IO [[String]]

Method to print the avg price closes, sorted by price and for all companies

estMoneyMade :: IConnection conn => conn -> Int -> IO [[String]]

Method to print the avg price closes, sorted by price and for all companies

avgPriceByMonth :: IConnection conn => conn -> Int -> IO [[String]]

Method to print the average prices for each month.

avgPriceByYear :: IConnection conn => conn -> IO [[String]]

Method to print the average prices for each year, for all companies ordered by AverageClose desc.

printString :: IConnection conn => conn -> IO String

Method to print error msg for companies

buildString :: [Company] -> String

Method to build a string with all the current companies in the db. To show the prompt

printpbc :: IConnection conn => conn -> Int -> IO ()

Method to display prices by company

printpbcdr :: IConnection conn => conn -> Int -> String -> String -> IO ()

Method to display prices by company

printcheapest :: IConnection conn => conn -> Int -> IO ()

Method to display cheapest prices for a company

printcheapestdr :: IConnection conn => conn -> Int -> String -> String -> IO ()

Method to display prices by company

printhighest :: IConnection conn => conn -> Int -> IO ()

Method to display highest prices for one company

printhighestdr :: IConnection conn => conn -> Int -> String -> String -> IO ()

Method to display prices by company

printhighestpriceall :: IConnection conn => conn -> IO ()

Method to display highest prices for all companies

printhighestalldr :: IConnection conn => conn -> String -> String -> IO ()

Method to display prices by company

printcheapestpriceall :: IConnection conn => conn -> IO ()

Method to display lowest prices for all companies

printcheapestpricealldr :: IConnection conn => conn -> String -> String -> IO ()

Method to display prices by company

printavgall :: IConnection conn => conn -> IO ()

Method to display avg prices for all companies

printavgpricedr :: IConnection conn => conn -> String -> String -> IO ()

Method to display prices by company

printestmoneymade :: IConnection conn => conn -> Int -> IO ()

Method to display the moneymade on each day for one company

printavgpricebymonth :: IConnection conn => conn -> Int -> IO ()

Method to the average close for one company, grouped by month

printavgpricebyyear :: IConnection conn => conn -> IO ()

Method to the average close for all companies, grouped by year