CSVHTTP

Description

Module that performs http requests as well as parsing.

Documentation

downloadURL2 :: String -> IO (Either String String)

This function will download a given url and return its content as a String This uses simpleHTTP which is not as fast as simpleHttp This function also uses the Either Monad for exceptions

csvFile :: ParsecT String u Identity [[[Char]]]

line :: ParsecT String u Identity [[Char]]

cell :: ParsecT String u Identity [Char]

quotedCell :: ParsecT String u Identity [Char]

quotedChar :: ParsecT String u Identity Char

eol :: ParsecT String u Identity String

parseURLs

Arguments

:: String
String representation of html file

-> [String]

Basic method for extracting all urls of a given text (String)

getLink

Arguments

:: [String]
String representation text from html file

-> String

This function returns a string which is the url to a .csv file

convertToStock

Arguments

:: [String]
Takes a list of Strings and returns a Stock data type

-> Stock

parseCSV :: String -> Either ParseError [[String]]

Parses the csv file and returns the file as a matrix wrapped by a monad

justTheList :: Either ParseError [[String]] -> [[String]]

Returns the file without the monad

removeHeader :: [[String]] -> [[String]]

Removes the table header from the file, we don't need this.

addKey :: [[String]] -> [String] -> [[String]]

appends a key to a list to represent a company key. This will be needed as a FK in the DB