Main

Documentation

main :: IO ()

The main function provides different functionalities:

addcomps

will provide a list of companies to be added into the db

updatecomp

will download a html file and parse it to get a new .csv file string to add to the db

@parsestock

Displays companies from the db, to the terminal

@showcompanies

Displays companies from the db, to the terminal

showstock

Displays all stock items for a given company

deletecompany

Removes a given company by it's ID

pricebycomp

Displays the Close, Date and Name for a given company

cheapestprice

Displays the cheapest close price for all companies

highestprice

Displays the highest close price for all companies

highestpriceall

Displays the highest close price for all companies

cheapestpriceall

Displays the cheapest close price for all companies

avg

Calcuates the avg price,for all companies

avgprice

Calculates the avg price by specifiyed year range, for all companies

estmoneymade

Calculates the estimated money made for a given comapny

avgpricebymonth

Calculates the avg price by month, group by year and company

avgpricebyyear

Calculates the avg price by year, group by year and company

getcsvlinks

Retrieves all csv links

getcsvlinks :: IO ()

addcomps :: IConnection conn => conn -> IO ()

updatecomp :: IConnection conn => conn -> String -> IO ()

deletecompany :: IConnection conn => conn -> String -> IO ()

prompt :: IConnection conn => conn -> IO ()

showcompanies :: IConnection conn => conn -> IO ()

showstock :: IConnection conn => conn -> String -> IO ()

parsestock :: IConnection conn => conn -> String -> IO ()

pricebycomp :: IConnection conn => conn -> String -> IO ()

pricebycompdr :: IConnection conn => conn -> String -> String -> String -> IO ()

cheapestprice :: IConnection conn => conn -> String -> IO ()

cheapestpricedr :: IConnection conn => conn -> String -> String -> String -> IO ()

highestprice :: IConnection conn => conn -> String -> IO ()

highestpricedr :: IConnection conn => conn -> String -> String -> String -> IO ()

cheapestpricealldr :: IConnection conn => conn -> String -> String -> IO ()

highestpriceall :: IConnection conn => conn -> IO ()

highestpricealldr :: IConnection conn => conn -> String -> String -> IO ()

avgall :: IConnection conn => conn -> IO ()

avgpricedr :: IConnection conn => conn -> String -> String -> IO ()

estmoneymade :: IConnection conn => conn -> String -> IO ()

avgpricebymonth :: IConnection conn => conn -> String -> IO ()

avgpricebyyear :: IConnection conn => conn -> IO ()

stocksToPull :: [[Char]]

csvlinksToList :: [String] -> [String]

csvlinksToCompany :: [String] -> [String] -> Int -> [Company]

csvlist :: [String]

companyList :: [Company]